+++ regattaserver/kv_test.go
@@ -10,6 +10,7 @@ import (
 	"github.com/jamf/regatta/regattapb"
 	"github.com/jamf/regatta/storage/errors"
 	"github.com/jamf/regatta/util/iter"
+	"github.com/lni/dragonboat/v4"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/mock"
 	"github.com/stretchr/testify/require"
@@ -100,7 +101,17 @@ func TestKVServer_RangeError(t *testing.T) {
 		Table: table1Name,
 		Key:   []byte("foo"),
 	})
-	r.EqualError(err, status.Errorf(codes.Internal, "unknown").Error())
+	r.EqualError(err, status.Errorf(codes.FailedPrecondition, "unknown").Error())
+
+	t.Log("Get retry-safe error")
+	storage = &mockKVService{}
+	storage.On("Range", mock.Anything, mock.Anything).Return((*regattapb.RangeResponse)(nil), dragonboat.ErrSystemBusy)
+	kv.Storage = storage
+	_, err = kv.Range(context.Background(), &regattapb.RangeRequest{
+		Table: table1Name,
+		Key:   []byte("foo"),
+	})
+	r.EqualError(err, status.Errorf(codes.Unavailable, dragonboat.ErrSystemBusy.Error()).Error())
 }
 
 func TestKVServer_RangeUnimplemented(t *testing.T) {
@@ -200,7 +211,7 @@ func TestKVServer_IterateRangeError(t *testing.T) {
 		Table: table1Name,
 		Key:   []byte("foo"),
 	}, srv)
-	r.EqualError(err, status.Errorf(codes.Internal, "unknown").Error())
+	r.EqualError(err, status.Errorf(codes.FailedPrecondition, "unknown").Error())
 
 	t.Log("Get unknown send error")
 	storage = &mockKVService{}
@@ -213,6 +224,16 @@ func TestKVServer_IterateRangeError(t *testing.T) {
 		Key:   []byte("foo"),
 	}, srv)
 	r.EqualError(err, status.Errorf(codes.Internal, "uknown send error").Error())
+
+	t.Log("Get retry-safe error")
+	storage = &mockKVService{}
+	storage.On("IterateRange", mock.Anything, mock.Anything).Return((iter.Seq[*regattapb.RangeResponse])(nil), dragonboat.ErrSystemBusy)
+	kv.Storage = storage
+	err = kv.IterateRange(&regattapb.RangeRequest{
+		Table: table1Name,
+		Key:   []byte("foo"),
+	}, srv)
+	r.EqualError(err, status.Errorf(codes.Unavailable, dragonboat.ErrSystemBusy.Error()).Error())
 }
 
 func TestKVServer_IterateRangeUnimplemented(t *testing.T) {
@@ -337,7 +358,17 @@ func TestKVServer_PutError(t *testing.T) {
 		Table: table1Name,
 		Key:   []byte("foo"),
 	})
-	r.EqualError(err, status.Errorf(codes.Internal, "unknown").Error())
+	r.EqualError(err, status.Errorf(codes.FailedPrecondition, "unknown").Error())
+
+	t.Log("Put retry-safe error")
+	storage = &mockKVService{}
+	storage.On("Put", mock.Anything, mock.Anything).Return((*regattapb.PutResponse)(nil), dragonboat.ErrSystemBusy)
+	kv.Storage = storage
+	_, err = kv.Put(context.Background(), &regattapb.PutRequest{
+		Table: table1Name,
+		Key:   []byte("foo"),
+	})
+	r.EqualError(err, status.Errorf(codes.Unavailable, dragonboat.ErrSystemBusy.Error()).Error())
 }
 
 func TestKVServer_Put(t *testing.T) {
@@ -398,7 +429,17 @@ func TestKVServer_DeleteRangeError(t *testing.T) {
 		Table: table1Name,
 		Key:   []byte("foo"),
 	})
-	r.EqualError(err, status.Errorf(codes.Internal, "unknown").Error())
+	r.EqualError(err, status.Errorf(codes.FailedPrecondition, "unknown").Error())
+
+	t.Log("Delete retry-safe error")
+	storage = &mockKVService{}
+	storage.On("Delete", mock.Anything, mock.Anything).Return((*regattapb.DeleteRangeResponse)(nil), dragonboat.ErrSystemBusy)
+	kv.Storage = storage
+	_, err = kv.DeleteRange(context.Background(), &regattapb.DeleteRangeRequest{
+		Table: table1Name,
+		Key:   []byte("foo"),
+	})
+	r.EqualError(err, status.Errorf(codes.Unavailable, dragonboat.ErrSystemBusy.Error()).Error())
 }
 
 func TestKVServer_DeleteRange(t *testing.T) {
@@ -448,7 +489,16 @@ func TestKVServer_TxnError(t *testing.T) {
 	_, err = kv.Txn(context.Background(), &regattapb.TxnRequest{
 		Table: table1Name,
 	})
-	r.EqualError(err, status.Errorf(codes.Internal, "unknown").Error())
+	r.EqualError(err, status.Errorf(codes.FailedPrecondition, "unknown").Error())
+
+	t.Log("Txn retry-safe error")
+	storage = &mockKVService{}
+	storage.On("Txn", mock.Anything, mock.Anything).Return((*regattapb.TxnResponse)(nil), dragonboat.ErrSystemBusy)
+	kv.Storage = storage
+	_, err = kv.Txn(context.Background(), &regattapb.TxnRequest{
+		Table: table1Name,
+	})
+	r.EqualError(err, status.Errorf(codes.Unavailable, dragonboat.ErrSystemBusy.Error()).Error())
 }
 
 func TestKVServer_Txn(t *testing.T) {